Using JSON Schema v1.3.1.
I think a bug might have been introduced in 1.3.2. An array_key_exists() error message is generated by the library when validating the configuration file against the library schema. Staying with 1.3.1 appears to be fine, but I opened an issue just in case it is a bug and it is affecting others as well:

https://github.com/justinrainbow/json-schema/issues/50