Fixes #75, where @git-version@ was not replaced in the output path.
The problem was that I am using @git-version@ in other parts of the Box project. So as part of the build process _for_ Box, I am replacing that placeholder with the actual version number. So, when other people use this feature, it's searching for Box's current version number to replace.