Fix the composer autoloader location for cgr
There is a tool called cgr (for composer global require) that
install every tool that needs to be globaly available in their own
custom folder. It does so that each tool dependecy don't infuences
other tools dependencies.

This means that the logic that was previously used by the autoloader
location code for box is broken. For some reason the code used to look
for the last autoloader in the folder structure ( the one closest to
the root folder ). I have no idea why it was doing so.

For cgr to work it needs to be the first one found that needs to be
used. I have changed the logic accordingly in this PR as well as
removed some dead code (an option that was never used).